requirements.txt
django
djangorestframework
watchdog
ultralytics
torch==2.3.1 
torchvision==0.18.1 
torchaudio==2.3.1
